Don't include android billing library if purchases are disabled
https://community.stencyl.com/index.php/topic,61840.0.html

The problem here is that "com.android.vending.BILLING" is declared by
the com.android.billingclient:billing:4.0.0 library automatically. We
can't include it if we don't want that permission to appear.

Regardless of that issue, we shouldn't be including all of this code
if it's not going to be used. For now, if purchases are disabled,
the source is stubbed out and the purchases dependency is no longer
included.