Update Google Play Billing to 8.0.0
The only notable difference here is that queryPurchaseHistoryAsync
is gone. I'm unsure what benefit it may have provided as it was, though,
since purchases have been restored at startup since the last major
update.

It may have allowed some limited consumables restoration, but if that's
true, that's completely gone now. The answer to that is probably cloud
saves.